energy-utility-platform: Added RabbitMQ consumer and Websocket

+++ src/main/java/com/example/dtos/DeviceDTO.java
@@ -1,6 +1,7 @@
 package com.example.dtos;
 
 
+import com.example.entities.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,6 +13,8 @@ public class DeviceDTO {
     private Long id;
     private String description;
     private String address;
-    private double maxConsumption;
+    private float maxConsumption;
+
+
 
 }
